requirements.txt
aiohttp==3.7.4.post0
alembic==1.13.1
astroid==3.1.0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
blinker==1.7.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.1.7
colorama==0.4.6
dill==0.3.8
discord.py==1.7.3
distlib==0.3.8
filelock==3.13.3
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==4.8.0
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.2
numpy==1.22.1
pandas==1.3.5
platformdirs==4.2.0
psycopg2==2.9.9
pylint==3.1.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==1.26.8
virtualenv==20.25.1
Werkzeug==3.0.1
yarl==1.7.2
